requirements.txt
Flask>=2.0
Flask-RESTful>=0.3.9
Flask-SQLAlchemy>=2.5.1
Flask-swagger==0.2.14
PyMySQL==1.1.1
pytest>=7.1.2
python-dotenv==1.1.1
setuptools==80.9.0
Werkzeug>=2.0
